packages/cdk/pyproject.toml
[project]
name = "cdk"
version = "0.1.0"
description = "infrastructure"
authors = [{name = "Second Brain Team"}]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "aws-cdk-lib>=2.0.0",
    "constructs>=10.0.0",
    "common",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.scripts]
build-layer = "python packages/cdk/build_layer.py"